last commit
The initial code had problems for updating patients using the stack(Last in, Last out) stack format. The new changes provided ensures that that functionality is possible. All updates are done from the top element of the stack and an element can only be deleted from the top element in the stack.
The when the maximum size is reached an element cannot be added to the already full stack, although updating and deleting processes are still possible .